+++ plugin/neovim_codex.lua
@@ -70,7 +70,3 @@ end, {})
 vim.api.nvim_create_user_command("CodexCaptureSelection", function()
   require("neovim_codex").capture_visual_selection()
 end, { range = true })
-
-vim.api.nvim_create_user_command("CodexCaptureBlock", function()
-  require("neovim_codex").capture_current_block()
-end, {})